refactor: rename call_desugar.rs to call_rewrite.rs, lower_comparisons to lower_comparison_operators
- call_desugar.rs → call_rewrite.rs: now only contains call rewriting
  logic, no more desugaring
- lower_comparisons → lower_comparison_operators: matches the specificity
  of sibling functions (lower_wide_int_match_patterns, lower_patterns)
